% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_reader}
\alias{hex_reader}
\title{Binary file reader}
\usage{
hex_reader(path)
}
\arguments{
\item{path}{File path.}
}
\value{
Interface to methods as list.
}
\description{
This function provides convenience methods for reading and parsing binary 
data.
}
