% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hexdecode}
\alias{hexdecode}
\title{Decode binary data}
\usage{
hexdecode(block, encoding, nomatch = NA, collapse = NULL)
}
\arguments{
\item{block}{Vector of type `raw`.}

\item{encoding}{`data.frame` or `list` of length 2. First item denotes byte,
second item denotes equivalent decoded character.}

\item{nomatch}{Character on to which bytes not on the character map should
be decoded.}

\item{collapse}{An optional character string to separate the results, as in
paste.}
}
\value{
A list of the same length as the number of splits of the byte
  sequence.
}
\description{
Decode a block of binary data based on a specified character encoding.
}
