% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexmode.R
\name{hexmode}
\alias{hexmode}
\title{Hexadecimal display}
\usage{
hexmode(length = 0)
}
\arguments{
\item{length}{A non-negative integer specifying the desire length. Double
values will be coerced to integer: supplying an argument of length other
than one is an error.}
}
\description{
Create and represent integers in hexadecimal format. This extends
`base::hexmode`'s functionality by defaulting to coercion to type `integer`
if no explicit S3 method for a class has been defined. Moreover, behavior is
different with regards to `NA` values in unsuccessful coercion of vectors of
types `double` and `character`.
}
