% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hexsplit}
\alias{hexsplit}
\title{Split the elements of a raw vector}
\usage{
hexsplit(block, split)
}
\arguments{
\item{block}{Vector of type `raw`.}

\item{split}{Sequence of bytes represented as a character string, eg.
`"\\x50"` or `"\\x50\\x51"`.}
}
\value{
A list of the same length as the number of splits of the byte
  sequence.
}
\description{
Split the elements of a raw vector into chunks according to the matches
to byte sequence `split` within them.
}
