% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hextoi}
\alias{hextoi}
\title{Convert hexadecimal string to integer}
\usage{
hextoi(x)
}
\arguments{
\item{x}{A character vector of hexadecimal strings (prefix `0x` optional).}
}
\value{
A numeric vector of corresponding decimal integers.
}
\description{
Utility function for converting hexadecimal string either with prefix `0x` 
(eg. `0x5b`) or without (eg. `5b`) into integers. If integers are given as
input, they are returned.
}
