% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{parse_hex_between}
\alias{parse_hex_between}
\title{Parse a block of data between addresses}
\usage{
parse_hex_between(path, start, end, split)
}
\arguments{
\item{path}{File path.}

\item{start}{Integer or character string representing a hexadecimal number
to set the initial position of binary file connection.}

\item{end}{Integer or character string representing a hexadecimal number that
signifies the address up to which the file is read.}

\item{split}{Sequence of bytes represented as a character string, eg.
`"\\x50"` or `"\\x50\\x51"`.}
}
\value{
A list of the same length as the number of splits of the byte
  sequence.
}
\description{
Read a block of binary data between two addresses and split it into chunks
according to the matches to byte sequence `split` within them.
}
