% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{parse_hex_block}
\alias{parse_hex_block}
\title{Parse a block of data of specified size}
\usage{
parse_hex_block(path, offset, size = 16, split)
}
\arguments{
\item{path}{File path.}

\item{offset}{Integer or character string representing a hexadecimal number
to set the initial position of binary file connection.}

\item{size}{Number of bytes.}

\item{split}{Sequence of bytes represented as a character string, eg.
`"\\x50"` or `"\\x50\\x51"`.}
}
\value{
A list of the same length as the number of splits of the byte
  sequence.
}
\description{
Read a block of binary data of a specified size starting at a given offset
and split it into chunks according to the matches to byte sequence `split`
within them.
}
