% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{read_hex_between}
\alias{read_hex_between}
\title{Read a block of data between addresses}
\usage{
read_hex_between(path, start, end)
}
\arguments{
\item{path}{File path.}

\item{start}{Integer or character string representing a hexadecimal number
to set the initial position of binary file connection.}

\item{end}{Integer or character string representing a hexadecimal number that
signifies the address up to which the file is read.}
}
\value{
Vector of type `raw`.
}
\description{
Read a block of binary data between two addresses.
}
