% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{read_hex_block}
\alias{read_hex_block}
\title{Read a block of data of specified size}
\usage{
read_hex_block(path, size = 16, offset)
}
\arguments{
\item{path}{File path.}

\item{size}{Number of bytes.}

\item{offset}{Integer or character string representing a hexadecimal number
to set the initial position of binary file connection.}
}
\value{
Vector of type `raw`.
}
\description{
Read a block of binary data of a specified size starting at a given offset.
}
