% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_reader}
\alias{hex_reader}
\title{Binary file reader}
\usage{
hex_reader(path, size = 16, offset = 0)
}
\arguments{
\item{path}{File path.}

\item{size}{Number of bytes.}

\item{offset}{Integer or character string representing a hexadecimal number
to set the initial position of binary file connection.}
}
\value{
Interface to methods as list.
}
\description{
`hex_reader()` provides convenience methods for reading and parsing binary
data.
}
