% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hexsplit}
\alias{hexsplit}
\title{Split the elements of a raw vector}
\usage{
hexsplit(x, split)
}
\arguments{
\item{x}{Vector of type `raw`.}

\item{split}{Sequence of bytes to use for splitting. Coerced by `as.raw` to a
`raw` vector if possible. Byte-sized `integer` and `hexmode` vectors are
valid.}
}
\value{
A list of the same length as the number of splits of the byte
  sequence.
}
\description{
Split the elements of a raw vector into chunks according to the matches
to byte sequence `split` within them.
}
